refactor(ci): move source-helpers to test/helpers/ subdirectory
Avoids tap picking up the helper as a test file when using test/*.js glob.
Reverts package.json back to the clean glob form.
